requirements.txt
altair==6.0.0
annotated-types==0.7.0
antiword==0.1.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
cobble==0.1.4
colorama==0.4.6
colorlog==6.10.1
defusedxml==0.7.1
distro==1.9.0
doc2txt==1.0.8
fast-langdetect==1.0.0
fasttext-predict==0.9.2.4
fonttools==4.61.1
fpdf2==2.8.6
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
mammoth==1.11.0
MarkupSafe==3.0.3
narwhals==2.17.0
numpy==2.4.2
olefile==0.47
openai==2.24.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
referencing==0.37.0
requests==2.32.5
robust-downloader==0.0.2
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.54.0
striprtf==0.0.29
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
